+++ typescript/packages/wallets/chromia/package.json
@@ -12,7 +12,7 @@
     "module": "./dist/index.mjs",
     "types": "./dist/index.d.ts",
     "dependencies": {
-        "@goat-sdk/core": "workspace:*",
+        "@goat-sdk/core-v2": "workspace:*",
         "@chromia/ft4": "1.0.1",
         "postchain-client": "1.20.1"
     },
