+++ typescript/packages/wallets/chromia/src/consts.ts
@@ -1,0 +1,5 @@
+export enum CHROMIA_MAINNET_BRID {
+    ECONOMY_CHAIN = "15C0CA99BEE60A3B23829968771C50E491BD00D2E3AE448580CD48A8D71E7BBA",
+}
+
+export const CHR_ASSET_ID = "5f16d1545a0881f971b164f1601cbbf51c29efd0633b2730da18c403c3b428b5";
