+++ typescript/packages/wallets/chromia/src/types/ChromiaTransaction.ts
@@ -1,0 +1,5 @@
+export type ChromiaTransaction = {
+    to: string;
+    assetId: string;
+    amount: string;
+};
